refactor(telegram): split callbacks.ts into callback modules
callbacks.ts was 639 lines — one callback_query:data listener with a big
prefix-dispatch over settings/pulse/effort/model. Split into a
callbacks/ folder:
  - shared   — editOrIgnoreSame + answerCallbackQuerySafe + deps type
  - settings — settings:* (effort/proactive + stale-picker acks)
  - pulse    — pulse:*
  - effort   — effort:*
  - model    — model:* menu/backend/browse controller

index.ts re-exports answerCallbackQuerySafe and rebuilds registerCallbacks
with the identical prefix dispatch + agent fall-through. Importer + test
updated. Verified: answer-callback-query-safe suite passes (7 tests).
