+++ src/frontend/telegram/index.ts
@@ -22,7 +22,7 @@ import {
 } from "./commands/index.js";
 import { setAccessControl } from "./handlers/index.js";
 import { registerMiddleware } from "./middleware.js";
-import { registerCallbacks } from "./callbacks.js";
+import { registerCallbacks } from "./callbacks/index.js";
 import { log, logError } from "../../util/log.js";
 
 // ── Frontend interface ──────────────────────────────────────────────────────
